Salvando alterações antes de voltar para branch menu

+++ menu/views.py
@@ -7,6 +7,7 @@ from utils.utils import calcula_valor_total_carrinho
 from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.core.paginator import Paginator
+from django.http import JsonResponse
 
 
 @login_required(login_url='clientes:login_cliente')
@@ -22,17 +23,17 @@ def index(request):
 
     usuario = get_object_or_404(CustomUser, username=request.user)
     carrinho_vinculado = get_object_or_404(Carrinho, cliente=usuario)
-    
-    #Incorporacao da view carrinho_detalhado
+
+    # Incorporacao da view carrinho_detalhado
     item_carrinho = ItemCarrinho.objects.filter(carrinho=carrinho_vinculado)
     subtotal = calcula_valor_total_carrinho(item_carrinho)
 
-    itens_carrinho = ItemCarrinho.objects.filter(carrinho=carrinho_vinculado).count()
+    qtd_itens_carrinho = ItemCarrinho.objects.filter(carrinho=carrinho_vinculado).count()
 
     context = {
         'page_obj': page_obj,
         'carrinho': carrinho_vinculado,
-        'qtd_item_carrinho': itens_carrinho,
+        'qtd_item_carrinho': qtd_itens_carrinho,
         'item_carrinho': item_carrinho,
         'produtos': produtos,
         'subtotal': subtotal,
@@ -45,7 +46,7 @@ def index(request):
 def buscar(request):
     usuario = get_object_or_404(CustomUser, username=request.user)
     carrinho_vinculado = get_object_or_404(Carrinho, cliente=usuario)
-    itens_carrinho = ItemCarrinho.objects.filter(carrinho=carrinho_vinculado).count()
+    qtd_itens_carrinho = ItemCarrinho.objects.filter(carrinho=carrinho_vinculado).count()
 
     query = request.GET.get("q", '').strip()
     if query == "":
@@ -63,7 +64,7 @@ def buscar(request):
     context = {
         'page_obj': page_obj,
         'carrinho': carrinho_vinculado,
-        'qtd_item_carrinho': itens_carrinho,
+        'qtd_item_carrinho': qtd_itens_carrinho,
     }
 
     return render(request, 'menu/index.html', context)
@@ -76,21 +77,46 @@ def adicionar_ao_carrinho(request, id):
     produto = get_object_or_404(Produto, id=id)
     carrinho_vinculado = get_object_or_404(Carrinho, cliente=usuario)
     consulta = ItemCarrinho.objects.filter(produto=produto, carrinho=carrinho_vinculado)
-    if consulta.first():
+    if consulta.exists():
         item = consulta.first()
         item.quantidade += 1
         item.save()
 
     else:
-        item_carrinho = ItemCarrinho.objects.create(
+        item = ItemCarrinho.objects.create(
             produto=produto,
             carrinho=carrinho_vinculado,
             quantidade=1,
         )
-        item_carrinho.save()
 
-    messages.info(request, f'{produto.nome} adicionado ao carrinho')
-    return redirect(request.META.get('HTTP_REFERER', 'menu:index'))
+    item_carrinho = ItemCarrinho.objects.filter(carrinho=carrinho_vinculado)
+    subtotal = calcula_valor_total_carrinho(item_carrinho)
+    qtd_itens_carrinho = ItemCarrinho.objects.filter(carrinho=carrinho_vinculado).count()
+
+    # messages.info(request, f'{produto.nome} adicionado ao carrinho')
+    # return redirect(request.META.get('HTTP_REFERER', 'menu:index'))
+    print(        {
+            "status": "success",
+            "message": f"{produto.nome} adicionado ao carrinho",
+            "quantidade": item.quantidade,
+            "total": item.get_total(),
+            "produto_id": item.produto.id,
+            "total_price": subtotal,
+        })
+
+    return JsonResponse(
+        {
+            "status": "success",
+            "message": f"{produto.nome} adicionado ao carrinho",
+            "quantidade": item.quantidade,
+            "total": item.get_total(),
+            "produto_id": item.produto.id,
+            'qtd_item_carrinho': qtd_itens_carrinho,
+            "total_price": subtotal,
+            "produto_nome": produto.nome,
+        }
+    )
+
 
 
 @login_required(login_url='clientes:login_cliente')
@@ -102,14 +128,45 @@ def remover_do_carrinho(request, id):
     consulta = ItemCarrinho.objects.filter(
         produto=produto, carrinho=carrinho_vinculado
     ).first()
-    if consulta.quantidade > 1:
-        consulta.quantidade -= 1
-        consulta.save()
-    elif consulta.quantidade == 1:
-        consulta.delete()
+    if consulta:
+        if consulta.quantidade > 1:
+            consulta.quantidade -= 1
+            consulta.save()
+            quantidade_atual = consulta.quantidade
+        elif consulta.quantidade == 1:
+            consulta.delete()
+            quantidade_atual = 0
+    else:
+        return JsonResponse(
+            {"status": "error", "message": "Produto não encontrado no carrinho"},
+            status=400,
+        )
 
-    messages.warning(request, f'{produto.nome} removido do carrinho')
-    return redirect(request.META.get('HTTP_REFERER', 'menu:index'))
+    # messages.warning(request, f'{produto.nome} removido do carrinho')
+    # return redirect(request.META.get('HTTP_REFERER', 'menu:index'))
+    item_carrinho = ItemCarrinho.objects.filter(carrinho=carrinho_vinculado)
+    subtotal = calcula_valor_total_carrinho(item_carrinho)
+    qtd_itens_carrinho = ItemCarrinho.objects.filter(carrinho=carrinho_vinculado).count()
+    
+    print(        {
+            "status": "success",
+            "message": f"{produto.nome} removido do carrinho",
+            "quantidade": quantidade_atual,
+            "total": consulta.get_total() if quantidade_atual > 0 else 0,
+            "produto_id": produto.id,
+            "total_price": subtotal,
+        })
+    return JsonResponse(
+        {
+            "status": "success",
+            "message": f"{produto.nome} removido do carrinho",
+            "quantidade": quantidade_atual,
+            "total": consulta.get_total() if quantidade_atual > 0 else 0,
+            'qtd_item_carrinho': qtd_itens_carrinho,
+            "produto_id": produto.id,
+            "total_price": subtotal,
+        }
+    )
 
 
 @login_required(login_url='clientes:login_cliente')
@@ -124,8 +181,22 @@ def excluir_do_carrinho(request, id):
     if consulta.quantidade >= 1:
         consulta.delete()
 
-    messages.error(request, f'{produto.nome} excluído do carrinho')
-    return redirect(request.META.get('HTTP_REFERER', 'menu:index'))
+    qtd_qtd_itens_carrinho = ItemCarrinho.objects.filter(carrinho=carrinho_vinculado).count()
+    # messages.error(request, f'{produto.nome} excluído do carrinho')
+    # return redirect(request.META.get('HTTP_REFERER', 'menu:index'))
+    item_carrinho = ItemCarrinho.objects.filter(carrinho=carrinho_vinculado)
+    subtotal = calcula_valor_total_carrinho(item_carrinho)
+
+    return JsonResponse(
+        {
+            "status": "success",
+            "message": f"{produto.nome} excluído do carrinho",
+            "quantidade": 0,
+            'qtd_item_carrinho': qtd_qtd_itens_carrinho,
+            "produto_id": produto.id,
+            "total_price": subtotal,
+        }
+    )
 
 
 @login_required(login_url='clientes:login_cliente')
